requirements.txt
appdirs==1.4.3
argcomplete==1.9.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
boto3==1.9.123
botocore==1.12.123
certifi==2019.3.9
cffi==1.12.2
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
cryptography==2.4.2
docutils==0.14
durationpy==0.5
Flask==1.0.2
fs==2.4.4
fs-s3fs==1.1.0
future==0.16.0
hjson==3.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
MarkupSafe==1.1.1
more-itertools==7.0.0
placebo==0.9.0
pluggy==0.9.0
py==1.8.0
pycparser==2.19
pytest==4.4.0
python-dateutil==2.6.1
python-dotenv==0.10.1
python-slugify==1.2.4
pytz==2018.9
PyYAML==5.1
requests==2.21.0
s3transfer==0.2.0
selenium==3.141.0
six==1.12.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.6
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.15.1
wsgi-request-logger==0.4.6
zappa==0.48.2
